src/main/java/org/prgrms/kdt/engine/customer/service/CustomerService.java
package org.prgrms.kdt.engine.customer.service;

import org.prgrms.kdt.engine.customer.domain.Customer;
import org.prgrms.kdt.engine.customer.repository.CustomerRepository;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@Service
public class CustomerService {
    private final CustomerRepository customerRepository;

    public CustomerService(CustomerRepository customerRepository) {
        this.customerRepository = customerRepository;
    }

    @Transactional
    public void createCustomers(List<Customer> customers) {
        customers.forEach(customerRepository::insert);
    }
}
